refactor: Fix bug in AboutMeComponent's ngOnInit method
This commit fixes a bug in the ngOnInit method of the AboutMeComponent. The closing parenthesis and semicolon were misplaced, causing a syntax error. The bug is resolved by moving the closing parenthesis and semicolon to the correct position. Additionally, a console.log statement is added to log the value of this.user for debugging purposes.