refactor: Update btn-create-group component to include group creation modal
This commit updates the btn-create-group component to include a group creation modal. The HTML template and TypeScript component are modified to add the modal functionality. The modal is triggered when the "Create group" button is clicked, and it allows the user to create a new group. This enhancement improves the user experience by providing a convenient way to create groups directly from the btn-create-group component.