refactor: Include user.isOwnProfile condition in ListGroupComponent HTML template
This commit adds the user.isOwnProfile condition to the ListGroupComponent HTML template. The condition is used to determine whether to display the list of groups. If the user is viewing their own profile, the list of groups is shown; otherwise, it is hidden. This enhancement improves the user experience by only showing relevant information based on the profile being viewed.