refactor: Update angular.json to include FontAwesome CSS
This commit updates the angular.json file to include the FontAwesome CSS file in the styles array. By adding the "node_modules/@fortawesome/fontawesome-free/css/all.min.css" file, we ensure that the necessary FontAwesome styles are applied to the application. This improvement enhances the visual appearance and user experience of the application.