refactor: Update updateUser function to accept FormData
The updateUser function in auth.service.ts has been updated to accept FormData instead of a user object. This change allows for more flexibility when updating user information, as FormData can handle different types of data, such as files. The update improves the functionality and usability of the updateUser feature.