refactor: Update UserDataService to use User interface
The UserDataService in user_data.service.ts has been updated to use the User interface from user.interfaces.ts. This change ensures that the changeUser and setUser methods accept and emit the correct type of data. Additionally, the isCurrentUser method has been added to verify if the current user is the owner of the profile. This update improves the code structure and type safety.